requirements.txt
Flask
gunicorn
python-telegram-bot
mysql-connector-python
